Manage listings page (user authorization)

+++ app/Http/Controllers/ListingController.php
@@ -50,10 +50,18 @@ class ListingController extends Controller
     }
 
     public function edit(Listing $listing){
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
         return view('listings.edit', ['listing' => $listing]);
     }
 
     public function update(Request $request, Listing $listing){
+
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -74,9 +82,19 @@ class ListingController extends Controller
     }
 
     public function destroy(Listing $listing){
+
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+
         $listing->delete();
 
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    public function manage(){
+
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 
 }
